Added (missing) board image.
I was referencing the board image in the CSS file, but for some reason
that goes beyonf me, I was not tracking the image.  So I incorrectly
pushed my commits without the board image but with a reference to it in
the CSS file.  (Shame on me, I know!)

This commit corrects that mistake.