Simplified the CSS file.
Unified the four colors.  Now the coins on the board and the coin icons
next to the players' names have exactly the same color.

Also tweaked the colors to be slightly darker than the their 'plain'
color so that they can be seen more easily when they have their color's
background.  Specifically, I replaced 0xFF by 0xE0, so e.g. red was
0xFF0000 and now it is 0xE00000.