Implemented a rotate function.
The rotate function rotates an ordered pair around the center of the
board, whose coordinates are (x_c, y_c), by an integer number of 90
degrees.

The 'absolute to coordinates' transformation (a function already
implemented) works only for green's coins.  Applying a rotation to those
coordinates, it is now possible to display any color's coins.