New function: player's index to color.
Defined a new function that, given a player's index, returns its
corresponding color as a lowercase string.

This is meant to be used to access HTML elements easily.  For example, a
players coin could have the id 'greencoin_1', so that element can be
accessed easily given the player's index and the player's coin index.